+++ src/gray/uint64.hpp
@@ -24,7 +24,7 @@ namespace color
 
     template< >
      struct container< ::color::category::gray_uint64 >
-      : public  ::color::_internal::utility::container::Unsigned< std::uint64_t, std::uint8_t,  unsigned,  1, 64 >
+      : public  ::color::_internal::utility::container::Unsigned< std::uint64_t, std::uint64_t, unsigned, 1, 64 >
       {
       };
